Dense Ore Cleansing of 2014
I have removed all redundant dense ore definitions, as well as removed any and all ores that spawn exclusively in their own dimension outside of the overworld.  This includes nether ores (excluding nether quartz), ender ores, and Metallurgy Fantasy ores, as they spawn in the Twilight Forest only.  This brings our dense ore count from the previous 89 to the current 28.  This should help with #15

This decision was made to reinforce the idea that these ores are supposed to be rare, and worth finding.  Considering also that along with AOBD's help, it's possible to grind up each of these ores and get double the output, if not up to 5x from Mekanism machines.  Coupled with a dense version, that means 15-20x the output from a single ore.  considering it takes 24 to craft a full set of armor, this is a bit overpowered to say the least.  These rare ores will stay rare, and will not be receiving any help from a dense version.