Major Mod Updating
This is a big one.  Over the weekend I updated quite a few mods, and
thus I can't simply commit each one individually.  I apologize that this
commit is as large as it is, but  bear in mind 80-90% of the changes are
due to simply adding/updating the mods.

To my knowledge, I did tweak the following:

- Tweaked oregen for Mekanism/Metallurgy Ores slightly
- Added Big Reactors
- Added NEI Plugins/Addons
- Added Galacticraft
- Added Magical Crops
- Added Tinker's Steelworks
- Added WAWLA
- Added Tinker's Mechworks

Please take a moment to glance over the commit and change anything you
deem incorrect.