Removed Unnecessary CoFH Oregen configs
Any ore that CoFH was generating in a way that Metallurgy can do on its
own, will now be generated by Metallurgy, not CoFH.  hopefully this
eases up on FPS hangups, and in general should be cleaner.

Also added in various gravel ores to normal gravel pockets.  Will
probably have to be tweaked.