Initial Commit for Ender Chest Tweaks
Holy cow was this a difficult one to get working.  For now, it'll be
defined explicitly, and the better, more programmatic way of doing it
will be left commented out.  If that way ever works, I'd prefer to use
the loop, rather than the mess it is now.

But hey, it works.