Add Additional Item Prefixes to AOBD
Because why not?  I'm thinking we can use minetweaker to add recipes for them, or even CoFH to generate them.

Tiny Dust for grinding up nuggets, or even as an additional output from processing (Especially the SAG Mill).

Nugget for things that don't already give them, might as well be complete.

Ore for those nether ores we'd like to add that don't have an overworld counterpart.