Integrating Tiny Dusts, Part 1
Severely WIP and untested.  This has an estimated chance to succeed of ~0%.  Part of #135

Once I write down what the individual ingots, dusts, etc. are, then the script has a better chance of working correctly.  Also this file REALLY needs to be refactored at some point... this commit alone pushed it up to 150% of it's original size.