Un-Nerf Alloy Blast Induction Smelter Requirements
So far all I can see that is unecessarily added to the Blast Induction Furnace is steel and bismuth bronze.  The former already has a blast furnace dedicated (high oven) and the latter is more of what I'd consider an early game alloy, not something that needs an blast furnace induction smelter to melt.