Extend Casting Tick Delay
Considering 20 is way2fast, these delays will be based off of the ratio that 2 seconds for a single item's worth should be sufficient.  Since redstone melts down to 100 mB, and a block is 9x that, the total time it will take is 18 seconds, or 360 ticks.  Same goes for glowstone; it melts down to 250 mB, and a block is 4x that, so a full block/bucket will take 8 seconds, or 160 ticks.