Fix GC Ilmenite Ore Dictionary Registry
It technically IS ilmenite, and not purely titanium.  Before GT was added, this was fine.  Now that GT adds actual titanium, as well as ilmenite support? might as well register this correctly.  It'll still process to titanium for the most part, but in a few situations it'll give other metals as well.