Disable Stronghold Generation in Caveworld Dim
The Deep Caveworld should be fine alongside those in the overworld, but the normal one as well?  That's a bit much to have at least 9 strongholds (if they are even limited in these dimensions) on one server.  If we can disable the generation of strongholds in the overworld, that'd be nice as well, as we can then have getting to one be a real challenge.