Temp Fix for Possible Issue
Translocator's diamond nugget is registered under a non-conventional name in the ore dictionary, diamondNugget.  Because of the recent removal of PoorOres, which registered their nugget correctly, it's possible having a null ore dictionary registry is causing issues here.

This is a temporary fix, as hopefully in the future we can refactor everything, and by the time this script gets loaded, nuggetDiamond will have items registered in it.