Change EnderIO Machine Recipes
They won't cost a resonant version of the machines they require, but they'll cost a redstone cell frame.  This will pose the choice of either upgrading the machine to the top tier version of it, or using it to craft the EnderIO version.

The Alloy Furnace also has had its recipe changed to 3 basic RF furnaces, to better suit it's role.