Tweak TiC Configs & Scripts
Mostly cleanup, but one thing to note is that I've lowered the ratios from 1,000,000:1,000,000:1,000,000 to 1000:1000:1 considering that they are multiplied by the ingot value in mB, as well as drive home the fact that these ratios are BAD, not just inefficient.  These should be fixed once ModTweaker fixes the alloy removal issue.