Disable Main Menu Mobs
Has caused my client to crash to desktop in the past, and if we add Custom Main Menu in the future, it may end up conflicting anyways.  Either way, they didn't add a whole lot aside from some silliness, so for now we can remove and see if things improve, and if we really want them back, we can just re-enable it.