Fix Deep Dark Generation Hanging
This at least stops the Deep Dark from hanging during world generation, locally.  Still untested on server, as well as the fact that not every desired ore is generating there yet.  If this fixes the problem, we'll have to re-introduce more ores, as well as decide how to generate them.