Update Harvest Level Display Texts
Adds the actual number into the tooltip so players can see whether or not adamantine is better than sanguinite or not, as it's not immediately apparent otherwise.  Also, since this means each entry has at least one space in it, I've wrapped each value in quotation marks to hopefully fix any issues the spaces may cause.