Rename Harvest Level Tiers
This should fit better.  Before, there were a few odd level names, such as steel and adamantine.  Now, there's 3 levels per tier/dimension, and for the most part, each tier represents a metal that is not an alloy and whose tools go up a harvest level.  The "main path", if you will.