Hopefully Fix TE4 Upgrade Recipes
getTag doesn't take any arguments, and returns an IData.  withTag takes an instance of an IData, and updateTag takes one as well.  What this does now, is grab the output, set it's NBT tag equal to the ones from the input, and then update the Level tag to the next tier up.