Re-Arrange Some Ore Generation
With two new dimensions being added, we want to spread out a bit of the magical ores.  Now, Adamantine, Astral Silver, Atlarus, Prometheum, and Rubracium will be found in the Witchery Spirit World; while Amber and Infuscolium will be found in the Tropicraft Dimension.  Certus Quartz has been moved to the Twilight Forest to mirror that of the Nether Quartz in the Nether, and Shadow Iron is now limited to the Boneyard biome in the Nether.

Also re-enabled the underfluid generation for various ores as it has been fixed in CoFHCore.