Use Workaround to Possibly Fix Dense Osmium Ore
I'm not sure why Yellorite Ore has always worked, but Osmium Ore has an issue being dense oreified.  This is caused by the texture of the ore being a different resolution from the underlying block (in this case, Osmium Ore is 32x32, and stone is 16x16).  Setting the underlying block to Mekanism's Copper or Tin Ore will set the resolution of the underlying block to 32x32 as well, and since the texture is basically identical except underneath the ore pixels, which would be different if it were stone anyways, means it should be able to be used as the underlying block.