Disable Some Incorrect or Unecessary Recipes
Some nuggets were able to be crafted into ores, aluminum was crafting into some GC block without a name, idk man....

Also disabled Metlalurgy's machines, as they're just unecessary with the rest of the pack's content.  If we decide we want them back in the future, we can always just add a recipe for them.