Set Thaumcraft Research Difficulty to Easy
The minigame just feels unnecessary, plus this way players need to spend the actual aspects, rather than cheesing it with a chain of low-end aspects.  Tilt the cost towards research points, rather than wasted time and effort.  The research itself isn't going to be the content's balancing point, the crafting recipe is.