Initial Commit to Rework Biomes & Worldgen
Mostly works, but also doesn't.  forest biomes are basically everywhere, and should be weighted better, but the swamps have been cleared off the shorelines, the mountains and alps exist in their own height restrictions, and the temperature ranges should be all aligned the way we want.  From here on out it should be filling out the biome groups and sub-biomes with appropriate weighting, as well as finessing the category suitability parameters.

Part of #231.