Even more Biome Shifting
Seriously, there's so much work done here, and I'm not entirely sure what state everything's at, at this point.  The Ice Cap seems to be ok, but probably needs tweaking.  The Mesas should be better this way, since I can't seem to get IRL mesas to generate in any sensible way, so this will have to do.  Oceans might need some tweaking but should be fine.

On the bright side, more biome groups have been added near the borders of the deserts and snowy areas to try and refine the borders' "shapes" a bit.  I'd like there to be a nice transition into the frozen/arid wastes of the poles/equator, rather than what it has been.