Reorganize Quartz Ore Mining Levels
Generic Quartz Ore will now be Copper, and Certus/Nether Quartz will be 2 above that, just like the rest of the Nether Ores (and the fact that Certus is equivalent to Nether Quartz).  This means Generic Quartz can be obtained with a Copper Pickaxe, so that players can make an AE2 Grindstone before having to make Bronze.

Fixes #256.