Disable Skeletons Spawning from Gravestones
It's a problem that seems to just get worse and worse, especially if a player dies when entering a dimension like the nether, causing them to just die more and more often when entering.

If we decide to turn them back on, we can, but for now (especially during testing) we can leave them off.