Set Rock Smasher to automatically double ores
Let's see, use the AE2 grindstone, which takes no power/upkeep (besides the wood for the crank) to get a 50% chance to double, or build a small steam setup to power some rock smashers to get a 25% chance to double?  Seems completely backwards, so we're going to allow these to just automatically double; seems nice for the early game, especially if you find a setup in a village.

This is even furthered by the fact that TiC smelteries and High Ovens don't automatically double anymore.